+++ app/controllers/user_controller.rb
@@ -1,18 +1,13 @@
 class UserController < ApplicationController
+  respond_to :html, :js
+
   def add
     @post = Post.new
   end
 
   def show
     @footer_off = true
-    @posts = Post.order(:updated_at).reverse_order
-    @posts = @posts.where("suit = true")
-    @posts = @posts.where("created_at < ?", Time.now)
-    @posts = @posts.paginate(page: params[:page], per_page: 2)
-    respond_to do |format|
-      format.html
-      format.js
-    end
+    @posts = Post.tencet_index.pagination(params[:page])
   end
 
   def about
